requirements.txt
uvicorn==0.21.1
sqlalchemy==1.4.41
fastapi==0.93.0
fastapi-utils==0.2.1
pydantic==1.10.7
passlib==1.7.4
sqlmodel==0.0.8
mongoengine==0.27.0
confluent-kafka==2.1.1
python-dotenv==1.0.0
python-jose==3.3.0
psycopg2==2.9.6
sqlalchemy-utils==0.41.1
celery==5.2.7
redis==4.5.4
httpx==0.24.0
tiktoken==0.3.3
langchain==0.0.188
openai==0.27.0
chromadb
ptvsd